% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_clust.R
\name{em_clust}
\alias{em_clust}
\title{Preform EM clustering on a dataset.}
\usage{
EM(dat, num)
}
\arguments{
\item{dat}{data that can be made into a matrix for em clustering}

\item{num}{the number of clusters}
}
\value{
A list containing means of the clusters, clustering vector, and variance.
}
\description{
Preform EM clustering on a dataset.
}
